requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
feedparser>=6.0.0
requests>=2.25.0
transformers>=4.15.0
torch>=1.9.0
python-dateutil>=2.8.0
beautifulsoup4>=4.9.0